chore: main script moved into main directory
